게임리프트 서버 데이터를 파일에서 읽어오는 부분 제거

+++ Assets/03.Scripts/Managers/InGameManager.cs
@@ -150,7 +150,7 @@ namespace ED
             auth.LocalNickName = userInfo.userNickName;
             auth.PlayerSessionId = matchInfo.PlayerGameSession;
             client.localPlayerId = userInfo.userID;
-            client.ConnectAsync(matchInfo.ServerAddress, matchInfo.Port);
+            client.ConnectAsync(matchInfo.ServerAddress, (ushort)matchInfo.Port);
         }
 
         async UniTask StartAIModeGame()
